# service/ClientService.js
'use strict';

const Client = require('../dal/ClientDal');



exports.createClient = function(body) {
  return new Promise(function(resolve, reject) {
   Client.createClient(body,(err, data) => {
      if (err || !data) {
        reject({
          resBody: {
            payload: null,
            msg: err
          }
        })
      } else {
        resolve({
          resBody: {
            payload: data,
            msg: "succeed msg",
            code: 200
          },
        })
      }
    })
  });
}



exports.deleteClient = function(id) {
  return new Promise(function(resolve, reject) {
   Client.deleteClient(id,(err, data) => {
      if (err || !data) {
        reject({
          resBody: {
            payload: null,
            msg: err
          }
        })
      } else {
        resolve({
          resBody: {
            payload: data,
            msg: "succeed msg",
            code: 200
          },
        })
      }
    })
  });
}



exports.listClients = function(matchingParams,rps,rpi) {
  return new Promise(function(resolve, reject) {
   Client.listClients(matchingParams,rps,rpi,(err, data) => {
      if (err || !data) {
        reject({
          resBody: {
            payload: null,
            msg: err
          }
        })
      } else {
        resolve({
          resBody: {
            payload: data,
            msg: "succeed msg",
            code: 200
          },
        })
      }
    })
  });
}



exports.updateClient = function(body) {
  return new Promise(function(resolve, reject) {
   Client.updateClient(body,(err, data) => {
      if (err || !data) {
        reject({
          resBody: {
            payload: null,
            msg: err
          }
        })
      } else {
        resolve({
          resBody: {
            payload: data,
            msg: "succeed msg",
            code: 200
          },
        })
      }
    })
  });
}